Split Design system into various modules
To start making the design system module more manageable, I figured we
could split it into separate modules per component.
