+++ web/design_system/ds_logo.re
@@ -1,0 +1,9 @@
+module S = ReactDOMRe.Style;
+
+let style =
+  S.make(~maxWidth="80px", ~padding="1rem", ~display="inline-block", ());
+
+[@react.component]
+let make = () => {
+  <img style src="./assets/logo.svg" />;
+};
