+++ web/design_system/ds_nav.re
@@ -1,0 +1,6 @@
+module S = ReactDOMRe.Style;
+
+[@react.component]
+let make = (~children) => {
+  <nav> children </nav>;
+};
