+++ web/lib/js/design_system/ds_background.bs.js
@@ -1,0 +1,32 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+var style = {
+  background: "radial-gradient(ellipse at bottom, #1b2735 0%, #090a0f 100%)",
+  bottom: "0",
+  height: "100%",
+  left: "0",
+  overflow: "hidden",
+  position: "fixed",
+  right: "0",
+  top: "0",
+  width: "100%",
+  zIndex: "-1"
+};
+
+function Ds_background(Props) {
+  return React.createElement("div", {
+              style: style
+            });
+}
+
+var S = 0;
+
+var make = Ds_background;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
