+++ web/lib/js/design_system/ds_footer.bs.js
@@ -1,0 +1,34 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+var Ds_link = require("./ds_link.bs.js");
+var Metadata = require("../metadata.bs.js");
+var Ds_container = require("./ds_container.bs.js");
+
+var style = {
+  height: "50px",
+  textAlign: "center"
+};
+
+function Ds_footer(Props) {
+  return React.createElement(Ds_container.make, {
+              style: style,
+              children: null
+            }, "Built with ☕️ and ❤️ at", React.createElement(Ds_link.make, {
+                  href: "https://src.technology",
+                  children: "SRC"
+                }), React.createElement("br", undefined), "Index built from ", React.createElement(Ds_link.make, {
+                  href: "https://github.com/reuniverse/reuniverse/tree/" + Metadata.git_commit,
+                  children: Metadata.git_commit
+                }), " at " + Metadata.build_time);
+}
+
+var S = 0;
+
+var make = Ds_footer;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
