+++ web/lib/js/design_system/ds_global.bs.js
@@ -1,0 +1,30 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+var ReactDOMRe = require("reason-react/lib/js/src/ReactDOMRe.js");
+
+var font_style = {
+  color: "white",
+  fontFamily: "Montserrat, Helvetica Neue, Helvetica, sans-serif",
+  fontSize: "1rem"
+};
+
+var style = ReactDOMRe.Style[/* combine */0](font_style, { });
+
+function Ds_global(Props) {
+  var children = Props.children;
+  return React.createElement("section", {
+              style: style
+            }, children);
+}
+
+var S = 0;
+
+var make = Ds_global;
+
+exports.S = S;
+exports.font_style = font_style;
+exports.style = style;
+exports.make = make;
+/* style Not a pure module */
