+++ web/lib/js/design_system/ds_label.bs.js
@@ -1,0 +1,32 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+var style = {
+  backgroundColor: "#232323",
+  color: "white",
+  fontFamily: "Helvetica Neue, Helvetica, sans-serif",
+  fontSize: "1em",
+  fontWeight: "500",
+  minWidth: "50px",
+  padding: "5px",
+  textTransform: "uppercase",
+  borderRadius: "5px"
+};
+
+function Ds_label(Props) {
+  var children = Props.children;
+  return React.createElement("span", {
+              style: style
+            }, children);
+}
+
+var S = 0;
+
+var make = Ds_label;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
