+++ web/lib/js/design_system/ds_link.bs.js
@@ -1,0 +1,30 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+var Ds_label = require("./ds_label.bs.js");
+
+var style = {
+  color: "white",
+  textDecoration: "none"
+};
+
+function Ds_link(Props) {
+  var href = Props.href;
+  var children = Props.children;
+  return React.createElement(Ds_label.make, {
+              children: React.createElement("a", {
+                    style: style,
+                    href: href
+                  }, children)
+            });
+}
+
+var S = 0;
+
+var make = Ds_link;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
