+++ web/lib/js/design_system/ds_loading.bs.js
@@ -1,0 +1,25 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+var style = {
+  margin: "2rem auto",
+  textAlign: "center",
+  width: "200px"
+};
+
+function Ds_loading(Props) {
+  return React.createElement("div", {
+              style: style
+            }, "Loading...");
+}
+
+var S = 0;
+
+var make = Ds_loading;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
