+++ web/lib/js/design_system/ds_logo.bs.js
@@ -1,0 +1,26 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+var style = {
+  display: "inline-block",
+  maxWidth: "80px",
+  padding: "1rem"
+};
+
+function Ds_logo(Props) {
+  return React.createElement("img", {
+              style: style,
+              src: "./assets/logo.svg"
+            });
+}
+
+var S = 0;
+
+var make = Ds_logo;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
