+++ web/lib/js/design_system/ds_nav.bs.js
@@ -1,0 +1,17 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+function Ds_nav(Props) {
+  var children = Props.children;
+  return React.createElement("nav", undefined, children);
+}
+
+var S = 0;
+
+var make = Ds_nav;
+
+exports.S = S;
+exports.make = make;
+/* react Not a pure module */
