+++ web/lib/js/design_system/ds_package_card.bs.js
@@ -1,0 +1,60 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var Infix = require("../utils/infix.bs.js");
+var React = require("react");
+var Package = require("../model/package.bs.js");
+var Ds_label = require("./ds_label.bs.js");
+
+var style = {
+  background: "#F9F9F9",
+  color: "black",
+  fontSize: "10px",
+  listStyle: "none",
+  padding: "0.5rem"
+};
+
+var name_style = {
+  color: "black",
+  display: "inline-block",
+  fontWeight: "500",
+  margin: "0 1em",
+  overflow: "hidden",
+  whiteSpace: "nowrap",
+  width: "25%",
+  textOverflow: "ellipsis"
+};
+
+var desc_style = {
+  color: "black",
+  display: "inline-block",
+  margin: "0 1em",
+  overflow: "hidden",
+  whiteSpace: "nowrap",
+  width: "60%",
+  textOverflow: "ellipsis"
+};
+
+function Ds_package_card(Props) {
+  var pkg = Props.pkg;
+  return React.createElement("li", {
+              style: style
+            }, React.createElement("span", {
+                  style: name_style
+                }, pkg[/* name */0]), React.createElement("span", {
+                  style: desc_style
+                }, Infix.$$Option[/* <|> */0](pkg[/* description */1], "no description")), React.createElement(Ds_label.make, {
+                  children: Package.Target[/* to_string */1](pkg[/* target */4])
+                }));
+}
+
+var S = 0;
+
+var make = Ds_package_card;
+
+exports.S = S;
+exports.style = style;
+exports.name_style = name_style;
+exports.desc_style = desc_style;
+exports.make = make;
+/* react Not a pure module */
