+++ web/lib/js/design_system/ds_search_input.bs.js
@@ -1,0 +1,38 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var Curry = require("bs-platform/lib/js/curry.js");
+var React = require("react");
+var Ds_global = require("./ds_global.bs.js");
+var ReactDOMRe = require("reason-react/lib/js/src/ReactDOMRe.js");
+
+var style = ReactDOMRe.Style[/* combine */0](Ds_global.font_style, {
+      background: "none",
+      border: "0",
+      float: "right",
+      margin: "0",
+      outline: "0",
+      padding: "1rem",
+      width: "200px"
+    });
+
+function Ds_search_input(Props) {
+  var onChange = Props.onChange;
+  var placeholder = Props.placeholder;
+  return React.createElement("input", {
+              style: style,
+              placeholder: placeholder,
+              onChange: (function (e) {
+                  return Curry._1(onChange, e.target.value);
+                })
+            });
+}
+
+var S = 0;
+
+var make = Ds_search_input;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* style Not a pure module */
