+++ web/lib/js/design_system/ds_tiny_title.bs.js
@@ -1,0 +1,25 @@
+// Generated by BUCKLESCRIPT VERSION 6.0.2, PLEASE EDIT WITH CARE
+'use strict';
+
+var React = require("react");
+
+var style = {
+  margin: "0",
+  padding: "0"
+};
+
+function Ds_tiny_title(Props) {
+  var children = Props.children;
+  return React.createElement("h5", {
+              style: style
+            }, children);
+}
+
+var S = 0;
+
+var make = Ds_tiny_title;
+
+exports.S = S;
+exports.style = style;
+exports.make = make;
+/* react Not a pure module */
