--- web/lib/model/index.d
@@ -1,1 +1,0 @@
-../model/index.cmj :  ../model/package.cmj ../model/package.cmi
