--- web/lib/model/model.d
@@ -1,1 +1,0 @@
-../model/model.cmj :  ../model/index.cmj ../model/index.cmi ../model/package.cmj ../model/package.cmi ../model/version.cmj ../model/version.cmi
