--- web/lib/model/package.d
@@ -1,1 +1,0 @@
-../model/package.cmj :  ../utils/utils.cmj ../utils/utils.cmi ../model/version.cmj ../model/version.cmi
